requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
PyYAML>=6.0

# Stable Diffusion 3.5
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
peft>=0.9.0

# Image I/O
rasterio>=1.3.0
Pillow>=9.5.0

# Metrics
lpips>=0.1.4

# Visualization
matplotlib>=3.7.0

# Evaluation (optional)
scikit-learn>=1.3.0